support/third-perty-library/flask/tencent-cos/requirements.txt
Flask==0.12.2
qcloud-cos-v4==0.0.22
